% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipper.R
\name{as.clipper}
\alias{as.clipper}
\title{As Clipper}
\usage{
as.clipper(object)
}
\arguments{
\item{object}{The \code{\link[=SAM-class]{SAM}} or \code{\link[=Smooth-class]{Smooth}} object to be coerced.}

\item{Class}{character. You should live it as default.}

\item{strict}{logical. You should live it as default.}
}
\value{
and S3 clipper
}
\description{
Coerce a \code{\link[=SAM-class]{SAM}} or \code{\link[=Smooth-class]{Smooth}} object to clipper. A function useful to call \code{\link{highlight_substance}} when clipping is not necessary.
}
\examples{
x <- tile_read(base::system.file("extdata/tile.bsp", package = "uFTIR"))
x <- tile_base_corr(x)
x <- wavealign(x, primpke)
x <- tile_sam(x)
x <- as.clipper(x)
}
