% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_sam.r
\name{matrix_sam}
\alias{matrix_sam}
\title{Perform SAM between to matrices}
\usage{
matrix_sam(x, y, x.wl, y.wl)
}
\arguments{
\item{x}{matrix}

\item{y}{matrix}

\item{x.wl}{wavenumbers for x}

\item{y.wl}{wavenumbers for y}
}
\value{
vector
}
\description{
Perform the spectral angle mapper algorithm between to matrices.
}
