% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_info.R
\name{mosaic_info}
\alias{mosaic_info}
\title{Mosaic Info}
\usage{
mosaic_info(dmtfile)
}
\arguments{
\item{dmtfile}{Path to the .dmt file.}
}
\value{
An object of class \code{\link[=SpectralInfo-class]{SpectralInfo}}.
}
\description{
The function is equivalent to \code{\link{tile_read}}. It makes an \code{\link[=SpectralInfo-class]{SpectralInfo}} object using the .dmt file. It does not load the measured spectra, and it was written for mosaic images only. See the page of \code{\link[=SpectralInfo-class]{SpectralInfo}} class for a full description of the returned object.
}
\examples{
x <- mosaic_info(base::system.file("extdata/mosaic.dmt", package = "uFTIR"))
}
\seealso{
\code{\link[=SpectralInfo-class]{SpectralInfo}}
}
