% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam_load.R
\name{sam_load}
\alias{sam_load}
\title{Sam load}
\usage{
sam_load(dmdfile)
}
\arguments{
\item{dmdfile}{a connection to the .bin or .ban file (yes, is true).}
}
\value{
An object of \code{\link[=SAM-class]{SAM}}
}
\description{
The inverse function for sam_write
}
