% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam_write.r
\name{sam_write}
\alias{sam_write}
\title{Write a sam object to disk}
\usage{
sam_write(tile, sam)
}
\arguments{
\item{tile}{The tile form which the sam originates}

\item{sam}{The sam that you want to write to disk}
}
\description{
A wrapper to the C++ mosaic_sam_write
}
