% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Smooth-class}
\alias{Smooth-class}
\title{Smooth}
\value{
An S4 object of class "Smooth.
}
\description{
An object of this class is returned by a call to \code{\link{smooth_sam}}. It holds a smoother version of a SAM object, keeping only the clusters slot and slicing it to a user-given number of slices. This class have methods to plot, summarize, and extract the readings of pixel matching a given SpectralReference substance. You can use the generics \code{\link[graphics]{plot}}, \code{\link[base]{summary}}, and \code{\link{get_profile}} but beware, you might have to provide extra arguments! (so, please don't see only the generic but the method for this class, see the section see also).
}
\section{Slots}{

\describe{
\item{\code{smooth}}{a cube -array- (even if only has 1 slice) that holds a smooth version of a SAM clusters slot.}
}}

\examples{
NULL
}
\seealso{
\code{\link{plot_tile}} \code{\link{summary_sam}} \code{\link{get_profile_tile}} \code{\link{get_profile_sinfo}}
}
