% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{SpectralPack-class}
\alias{SpectralPack-class}
\title{Spectral Pack}
\value{
An S4 object of class "SpectralPack" ready for spectral angle mapping.
}
\description{
A call to the function \code{\link{wavealign}} (specifically a call to \code{\link{TileRead.wavealign}}) returns an object of class \code{\link[=SpectralPack-class]{SpectralPack}}. This class reunites the uFTIR Microscope readings loaded to R by a call to \code{\link{tile_read}} with a tarjet spectral library of class \code{\link[=SpectralReference-class]{SpectralReference}}. Since this object is an output of a call to \code{\link{wavealign}} both the \code{\link[=Tile-class]{Tile}} object and \code{\link[=SpectralReference-class]{SpectralReference}} have congruent wavenumbers. See \code{\link{wavealign}} how the resampling is done when calling different methods. A \code{\link[=SpectralPack-class]{SpectralPack}} object is used as imput when calling \code{\link{tile_sam}} to run the Spectral Angle Mapping algorithm.
}
\section{Slots}{

\describe{
\item{\code{Readings}}{an object of class \code{\link[=Tile-class]{Tile}}.}

\item{\code{Reference}}{an object of class \code{\link[=SpectralReference-class]{SpectralReference}}.}
}}

\examples{
NULL
}
