% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_base_corr.R
\name{tile_base_corr}
\alias{tile_base_corr}
\title{Baseline correction}
\usage{
tile_base_corr(data)
}
\arguments{
\item{data}{Object of class \code{\link[=Tile-class]{Tile}}.}
}
\value{
An object of class \code{\link[=Tile-class]{Tile}} congruent with data, with the spectra modified.
}
\description{
A function to substract the minimun of a pixel (vector) to have a minimal signal = 0, avoiding negative readings. Currently the only preprocessing method implemented.
}
\examples{
x <- tile_read(base::system.file("extdata/tile.bsp", package = "uFTIR"))
x <- tile_base_corr(x)
}
